Add themis object to project database
Uses hermes and some other logic to incorporate the data of a themis object into the athena database.

Three datasets created:
CellData, a table of all cell runs added and their stim params.
TunData, a concatenation of all the tuning data from each cell-run
Data, the join of CellData and TunData, where the join is performed from scratch each time a themis object is added (hopefully not too slow!)

Also logic for initialising the CellData and TunData when not extant yet.

And logic for checking whether the cell-run has been added already both in the hermes appendDataSet function and in simple assert lines