Fix insert cond_type, done at init and append level in hermes
This is a little dirty ... hard coded in as an update to the cell_data
that is passed in to the init and append functions.

As a result, there is no checking on the value for any redundancy
as there is with the cell data params.  IDEALLY, this would get updated
so that cond_type was part of cell data.