Add functionality for reloading raw_data files
When pickling Heph files, the self.Dat attrbute is removed due to its
size.

As a result, accessing the raw data, or using any functions that rely
on its availability, such as self.quickview(), is not possible.

The functionality added here uses the recorded path of the raw data file
to load the file again.  This loading is predicated on a flag in the
load function.

Loading is done according to the parameters used when initially loaded.
A function has been added for recording these parameters.

If no parameters were recorded, each of the loading functions are tried
with default parameters.